Move ZookeeperDependency to top level.
