+++ spring-cloud-zookeeper-discovery/src/test/groovy/org/springframework/cloud/zookeeper/discovery/ZookeeperDiscoveryISpec.groovy
@@ -27,8 +27,8 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient
 import org.springframework.cloud.client.loadbalancer.LoadBalanced
 import org.springframework.cloud.netflix.feign.EnableFeignClients
 import org.springframework.cloud.netflix.feign.FeignClient
-import org.springframework.cloud.zookeeper.common.CommonTestConfig
-import org.springframework.cloud.zookeeper.common.TestRibbonClient
+import org.springframework.cloud.zookeeper.discovery.test.CommonTestConfig
+import org.springframework.cloud.zookeeper.discovery.test.TestRibbonClient
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
