+++ spring-cloud-zookeeper-discovery/src/test/groovy/org/springframework/cloud/zookeeper/discovery/test/CommonTestConfig.groovy
@@ -13,25 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.zookeeper.common
+package org.springframework.cloud.zookeeper.discovery.test
 
 import groovy.transform.CompileStatic
-import org.springframework.web.client.RestTemplate
+import groovy.transform.PackageScope
+import org.apache.curator.test.TestingServer
+import org.springframework.cloud.zookeeper.ZookeeperProperties
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.util.SocketUtils
 
+@PackageScope
 @CompileStatic
-class TestServiceRestClient {
-
-	final RestTemplate restTemplate;
-
-	TestServiceRestClient(RestTemplate restTemplate) {
-		this.restTemplate = restTemplate
+@Configuration
+class CommonTestConfig {
+	@Bean(destroyMethod = 'close')
+	TestingServer testingServer() {
+		return new TestingServer(SocketUtils.findAvailableTcpPort())
 	}
 
-	String callService(String alias, String endpoint) {
-		return restTemplate.getForObject("http://$alias/$endpoint", String)
-	}
-
-	String callOnUrl(String url, String endpoint) {
-		return new RestTemplate().getForObject("http://$url/$endpoint", String)
+	@Bean ZookeeperProperties zookeeperProperties() {
+		return new ZookeeperProperties(connectString: "localhost:${testingServer().port}")
 	}
 }
