+++ spring-cloud-zookeeper-discovery/src/test/groovy/org/springframework/cloud/zookeeper/discovery/test/TestRibbonClient.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.zookeeper.common
+package org.springframework.cloud.zookeeper.discovery.test
 
 import org.springframework.web.client.RestTemplate
